ME-2560 - Updated the cropped image size on all apps

+++ source/api/models/Template.js
@@ -41,10 +41,10 @@ module.exports = {
             type: 'string'
         },
         secondNaviAspectRatio : {
-            type: 'integer'
+            type: 'float'
         },
         thirdNaviAspectRatio : {
-            type: 'integer'
+            type: 'float'
         },
     }
     /*,
